Refactor OCR and preprocessing logic into shared module
- Extracted duplicated OCR and preprocessing code into utils module
- Reduced code duplication across multiple app files
- Improved maintainability and consistency
